Moved server code in Main to it own class. Main now simply instantiates our controller and thus our application.
